anymals/requirements.txt
Django==4.1.7
six==1.16.0
django-bootstrap4==22.3
Pillow==9.2.0

uuid==1.30
django-filter==22.1